stored MongoDB cred in aws secret manager
