+++ get_cert.py
@@ -7,7 +7,35 @@ import boto3
 from botocore.exceptions import ClientError
 
 
-def get_secret():
+def get_privkey():
+
+    secret_name = "prod/flaskapp"
+    region_name = "us-east-1"
+
+    # Create a Secrets Manager client
+    session = boto3.session.Session()
+    client = session.client(
+        service_name='secretsmanager',
+        region_name=region_name
+    )
+
+    try:
+        get_secret_value_response = client.get_secret_value(
+            SecretId=secret_name
+        )
+    except ClientError as e:
+        # For a list of exceptions thrown, see
+        # https://docs.aws.amazon.com/secretsmanager/latest/apireference/API_GetSecretValue.html
+        raise e
+
+    secret = get_secret_value_response['SecretString']
+
+    # Your code goes here.
+    f = open("./test_cert/privkey.pem", "w")
+    f.write(secret)
+    f.close()
+
+def get_fullchain():
 
     secret_name = "prod/flaskapp2"
     region_name = "us-east-1"
@@ -34,4 +62,6 @@ def get_secret():
     f = open("./test_cert/fullchain.pem", "w")
     f.write(secret)
     f.close()
-get_secret()
+
+get_privkey()
+get_fullchain()
